requirements.txt
numpy>=1.23
pandas>=2.0.2
scikit-learn>=1.2
matplotlib>=3.6
seaborn>=0.12
netCDF4>=1.6
h5py>=3.9
rasterio>=1.3.8
cartopy>=0.21
torchinfo>=1.8
torchgeo>=0.4.1
tensorboard>=2.13
onnx>=1.15
tqdm>=4.65
easydict>=1.9
imageio>=2.31
utm>=0.7
psutil>=5.9
